Restructured the controller hierarchy. Added a ControllerBase which contains most of the common stuff between controllers. Edited the redirects and posts to their associated controller actions.
