+++ AutomobileRepairShop/Views/Shared/AddEmployee.cshtml
@@ -10,7 +10,7 @@
                     <div class="d-flex align-items-center h-custom-2 px-5 ms-xl-4 mt-5 pt-5 pt-xl-0 mt-xl-n5" id="employeeForm">
 
                         <!-- registration form -->
-                        <form asp-controller="Home" asp-action="AddEmployee" method="post" style="width: 23rem;">
+                        <form asp-controller="Admin" asp-action="AddEmployee" method="post" style="width: 23rem;">
 
                             <!-- "Register" title -->
                             <h3 class="fw-normal mb-3 pb-3" style="letter-spacing: 1px;">Add employee</h3>
