+++ AutomobileRepairShop/Views/Shared/EditAccounts.cshtml
@@ -73,7 +73,7 @@
                                 <div class="modal-body">
 
                                     <!-- edit form -->
-                                    <form id="editform" asp-controller="Home" asp-action="EditAccount" method="post" style="width: 23rem;">
+                                    <form id="editform" asp-controller="Admin" asp-action="EditAccount" method="post" style="width: 23rem;">
 
                                         <div class="form-outline mb-4">
                                             <input name="Id" type="name" for="id"
@@ -137,7 +137,7 @@
                         <div class="modal-dialog">
                             <div class="modal-content">
                                 <div class="modal-header">
-                                    <form id="deleteform" asp-controller="Home" asp-action="DeleteAccount" method="post" class="form-outline mb-4">
+                                    <form id="deleteform" asp-controller="Admin" asp-action="DeleteAccount" method="post" class="form-outline mb-4">
                                         <input name="Id" type="name" for="id"
                                                class="form-control form-control-lg" />
                                         <label class="form-label" for="id">User Id</label>
